python-parser-service/requirements.txt
# gRPC dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# PDF parsing
PyPDF2==3.0.1
pdfplumber==0.9.0

# DOCX parsing
python-docx==0.8.11

# Additional utilities
typing-extensions==4.8.0